Move filesize_formatting to utils for sharing
